Describe the steps required to add collaborator
In order to deploy into an application that exists
in the shared MoJ Heroku account, you need to use
the Heroku GUI (as the shared account user) and
add a personal account as a collaborator.

Ideally, this process should be scriptable using
either terraform or the heroku command-line client
But, I wasn't able to figure out if that was
possible, so I'm documenting the manual process,
for now.